.requirements.txt
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
fpdf==1.7.2
idna==3.10
msal==1.31.1
pycparser==2.22
PyJWT==2.10.1
PyYAML==6.0.2
requests==2.32.3
urllib3==2.3.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
fpdf==1.7.2
idna==3.10
msal==1.31.1
pycparser==2.22
PyJWT==2.10.1
PyYAML==6.0.2
requests==2.32.3
urllib3==2.3.0
